+++ scratch.js
@@ -32,3 +32,4 @@ let arr = [
 // let doc = nlp.fromSpec(arr[0]).debug()
 
 nlp('john smith and Sally Morris lives in new york').redact().debug()
+