+++ changelog.md
@@ -10,10 +10,12 @@ While all _Major_ releases should be reviewed, our only _large_ releases are **v
 
 #### 14.16.0 [June 2026]
 - **[new]** - apply redaction to #Organization, #Address, etc tags in .redact() method
+- **[new]** - apply #Redacted tag
 - **[new]** - .addresses() method
 - **[new]** - .out('spec') format
 - **[change]** - .money() selector and methods
 - **[fix]** - tokenize phone-numbers with parentheses
+- **[fix]** - replaceWith keep-tag issue
 - **[update]** - dependencies
 
 #### 14.15.1 [May 2026]