+++ changelog.md
@@ -9,7 +9,7 @@ While all _Major_ releases should be reviewed, our only _large_ releases are **v
 
 
 #### 14.16.0 [June 2026]
-- **[new]** - apply redaction to #Organization, #Address, etc tags in .redact() method
+- **[new]** - apply redaction to #Organization, #Address, #Url etc tags in .redact() method
 - **[new]** - apply #Redacted tag
 - **[new]** - .addresses() method
 - **[new]** - .out('spec') format