+++ changelog.md
@@ -8,7 +8,7 @@ While all _Major_ releases should be reviewed, our only _large_ releases are **v
 
 
 
-#### 14.16.0 [June 2026]
+#### 14.16.0 [July 2026]
 - **[new]** - apply redaction to #Organization, #Address, #Url etc tags in .redact() method
 - **[new]** - apply #Redacted tag
 - **[change]** - .redact() method now untags specific tags when redacting